Add libminijail
Add libminijail and a build setup. In the upstream libminijail
the build of the lib itself is done by the chrome or android
build system. For north a custom `build.rs` is added which compiles
the C files. The syscall table generation is not yet ported and
committed statically.

For building the minijail crates on non linux systems a mock that
spawns a plain process and ignores most of the linux features is
added.