Retry on EAGAIN for getting loop device status
The ioctl call to loop-control with LOOP_SET_STATUS64 can potentially
fail with an EAGAIN errno code. This error can be treated by retrying the
call after some time.

This change retries the ioctl call if EAGAIN is received with a delay
interval of 50 milliseconds between calls for a maximum of 3 retries
before resulting in an error.