Installing npk by streaming the file
A file can now be streamed to the north runtime. While streaming,
no other commands can be sent to the runtime.
Once the file is completely transmitted, it will be installed if
possible. A confirmation responds will indicate the success.
To install npk with the correct name, we expect an Installation message
before the actual data stream that informs north about the size of
the file and the name of it.

- Decouple manifest extraction from installation
- extract manifest without installing
- Introduce error handling for installation
- prevent installing an application id that is already installed
- map installation problems to api errors
- port new error types to mac mock
- separate error types into its own module
- add error types for verity header problems
- make constructor of Version usable outside