Use mounts from manifest
Specific either bind mounts or data mounts in the manifest.
Bind mounts are mounted as before. The data mounts in the
mount list in the manifest is a replacement for the default
data dir. The global data dir option is obsolete because this can
be configured via a custom bind mount.