Fix hanging stop calls in integration tests
The integration tests spawn `nstar` to stop containers. If the containers
were not running, `nstar` would not issue any stop command. The tests were
hanging waiting for the message of "Stopped [container name]" in the
runtime log.

This change removes that assertion. Also adds a `build.rs` to `tests` to
build the `test_container` using `sextant`.