Refactor target architecture check
Filtering *all* files in the container dirs is lazy. Normally there
should be only npks that can be actually executed by north that is
pointed to this dir. Instead of checking the `arch` tag in the manifest
configure a platform tag in the north configuration in order to runtime
check if the npk is suitable to run on this instance. The platform tag
in the manifest is currently permissive. Same for the one in the north
configuration.

Improve the building of the examples by not iterating of a set of
architectures but provide a explicit one if cross compilation is needed.
Otherwise build for the host. The architecture iteration feels more like
a CI/check thing which is not the intention of the examples.