feat: examples: Verify that build script is invoked from top-level dir
This patch adds a check for the invocation point from which
the build script was called and bails early with a message
when such a point is not the top-level directory.

This assumes the existence of the `git` app on the machine.
Alternatively, we could check solely based on `$(basename $0)`,
however that would not work for cases where the top-level directory
differs from `northstar`.