Set PR_SET_PDEATHSIG inside the application process
Setting PR_SET_PDEATHSIG causes the kernel to signal the
application process when the parent (runtime) exits. This
should normally be done via the minijails hooks which are C
function pointes that can be registered. Enhancing the minijail
Rust wrapper to pass this sort of data to minijail is way
more invasive regarding future updates than this one prctl.