Refactor manifest mounts
The mount flags on a persist mount do not make sense since this
shall always be rw. Change this to

```
/foo: persist
```

A mount of /dev can only be on /dev and is default with a minimal
set of devices (done by minijail). Optionally a mount entry

```
/dev: full
```

does a full fledged rw bind mount of `/dev` if more than the minimal
set is needed.