minijail: add support for capabilties and groups
This adds two new optional fields to the manifest:
  capability_str
  supplgroups_str

The capability_str is a list of capabilities in the form 'cap_xxx'
as per capabilities(7)
  cap_net_raw cap_chown cap_mknod
This list is then converted to a capabilities vector ; the spawned
container will then have these capabilities. The default is still
0 (no capabilities). The current set can be viewed in /proc/self/status

The supplgroups_str is a list of supplementary group names that the
spawned container should be a part of. For example, on Android
  adb inet
Only names are allowed, not the symbolic number reprentation.

This list is converted into the actual gid list which minijail will
use in a setgroups(2) call.

The added functionality is in a new file to avoid polluting the already
bloated minijail/libminijail.c