firecracker: add to minijail for launching VM
This adds support to minijail for starting a executable
that will launch a VM to run the executable from a container
instead of directly launching the container.

The linkage between the northstar Rust code and Minijail is
less than optimal.

There are three phases to launching a VM:
	setup: create config files are setup bind mounts
	launch: excve() a launch executable instead of the target
		/init that is specified in the container. This
		launch executable will then launch firecracker
	execute: configure the VM based on the config files and
		launch the application from the filesystem
		mounted from the block device on host.

The required resource container and test program (another container)
are in a different commit.