firecracker: add support files for resource container
This adds the directory structure for the support files
that are required for creating a VM resource container.
It also adds the scripts that are necesssary both in the
setup and launch phase of the VM.

A subsequent commit will checkin an example resource
container for aarch64.

There is a script in firecracker/tools/bin/fc_mkinitrd
that shows how to build the initrd for the resource container.

Each of the required tools has a README that describes where
the source comes from, and any patches that are required.

There is a PDF that describes the process for launching a VM
in firecracker/doc/northcracker.pdf