minijail: use pid namespace
There was a concern about a deadlock in minijail when using
pid namespaces, so they were disabled. This had the side-effect
of not cleaning up all processes when the container was stopped.

Northstar already enables run_as_init and the main minijail concern
is when that option is not used. Testing shows that with northstar,
even having minijail do a wait (run_as_init == 0), no deadlocks were
observed.

I believe that since the minijail code is called in a dedicated forked
async context from the rust code, the risk of deadlocking the parent
(as described by the minijail comment in minijail_run_internal() ) is
not present.

So, the solution is to enable pid namespaces in Process() in
runtime/minijail.rs (actually just remove the comment that disables them).

This means that if a container spawns multiple processes, they will
all be cleaned up when the container is stopped.

The alternative would be to have the forked child process in minijail
become a process group leader, and then use the kill(-pid, signal)
mechanism to kill all processes in that process group. This works fine,
but runs the risk if a process in the container switches process groups,
then it would not work.