Simplification of builtin container
The cargo build run from the build.rs is weird. The cross compilation
use case and the race around the build dir from the parallel running
cargo instances.
Replace the binary with an `echo` invocation. Dump the manifest during
build time instead of maintaining it in a file.

The tempdir which is the internal repository must be kept in `State`,
otherwise it get's removed.