Add debug facilities: strace, perf, log and ns
The runtime allows to be configured to automatically attach
a `strace` or `perf` instance to started processes and dump
their output to `log_dir`. A option in the runtime configuration
allows to disable the runtimes mount namespace for container
inspection. The stdout/stderr of applications can be configurated
to go to be discarded, go the runtimes log or reused the runtimes
stdout/stderr.