Refactor container mounts and organization
Instead of iterating all containers at startup and
mount sequentially the mounting is done lazy when a container
is about to be started. A container remains mounted when
it's stopped. All needed resources that a container needs are
mountend when a container is started. If one of the resources
fails to mount the start of the container failes.

Every container is indentified by a tuple with name, version and
repository. Resources also include the repository which is a
subject to be discussed. Multiple continers with the same name
are possible within differnet repositories.