namespaces: netlink sockets for network namespace
This uses netlink sockets instead of the IP utility to
create the devices within a namespace. Doing so avoids calls
to fork/exec for the 'ip' utility, in addition to removing
the necessity for bind mounts that the 'ip' utility uses to
maintain namespace persistence.

As a result, the time to create the namespace and the devices
inside it goes from about 800ms to roughly 150ms.

The bind mounts are no longer needed because the namespace is
private to a single process, and can therefore be referenced via
a simple file descriptor.

The API for the netlink sockets was determined by strace and
gdb of the 'ip' utility.