Handle new connections during shutdown
Currently it is possible to accept new connections during a shutdown.
This causes a sort of dead lock where the shutdown cannot terminate as
log as connections remain open but requests cannot be answer because the
runtime is just waiting for the shutdown.

To handle this situation, if a shutdown request is issued, the console
is closed for new connections and request in the event loop are
ignored.