minijail: re-enable pid namespaces
This uses a consistent interface to fork/clone in minijail. Earlier
versions mixed library calls (which acquire futexes during atfork()
calls) with directly invoked syscalls, bypassing the library handling.
As documented in the minijail code, this can lead to deadlocks. Most
likely, when first developed, the library code was not available.

With this fix, and the recent fixes to the runtime for handling
process termination, a 'spawner' test of 20 clients starting and
stopping can run in the release build for more than 36 hours.