northstar: Implement Codec in terms of tokio_util::codec::LinesCodec
Instead of framing new lines directly and then the dealing
with underlying JSONs, we leverage tokio_util::codec::LinesCodec
to frame input for us according to new lines,
leaving us with the single task of dealing with JSONs.

Notice that LinesCodec may return a MaxLineLengthExceeded
when a user sets such limit. Even if we don't
limit the maximum length, we cannot prove that to the
compiler and therefore we need a little map_err to line types up.