northstar: Add a simple test to check Codec consistency
Given that decoding and encoding are related by
the abstract property (omitting details for brevity):

`forall m in Message. decode(encode(m)) == m`

We check such property by leveraging proptest.

Notice that the `mk_simple_payload` is largely
simplified in the sense that we don't actually
consider all the possible variants, which might
be good enough as a starting point,
since we aren't particularly interested in
the JSON unmarshaling/marshalling,
but rather in the consistency of decoder/encoder.

Future patches might want to extend
`mk_simple_payload` for completeness.