Check stop token while waiting for request
Check the stop token while waiting for a request to be
processed. If the token is canceled in between the connection
is closed.

Fix the tmpfile writing. Using the filename of a NamedTmpFile
without holding the struct will result in the removal of the
file. The file created with the name of tmpfile is never removed.