Use island instead of minijail if both enabled
These two features seem to be mutually exclusive. When the project is
build, run or tested with the `--all-features` flag, the RuntimeLauncher
will be defined twice. This change adds an extra check in case the two
features are enabled that will cause the definition of RuntimeLauncher
that uses island to be prefered over minijail.