Move run_dir bind mount to main
The self mounted run_dir is something integration specific and therfore
should not be done withing the runtime lib. When running northstar for
example as an android service all this ceremony is not needed and done
by Android init.

Move responsibility of the configured directories also out of the lib
into it's user but keep the checks whether the dirs are read and
writeable for the runtime.

Dump backtraces on panics in tests.