Add relaxed and critical autostart variants
If a container is marked with Autostart::Critical, the runtime exits
with a non zero exit code when the container isn't mountable on start or
the container exits at any point in time with a non zero exit code.

The Autostart::Relaxed setting attempts to start the container but just
warns about any failure during the mount or start.