Refactor client api and remove test-container datadir
Passing the test instructions to the test-container via a file is weird
because the test fixture needs to know the exact path of the rw data dir
of the test container. Add a possibility to add additional command line
arguments and env variables to the `Start` command.

Update the client api to accept a `TryInto<Container>` everywhere where a
container name and version is needed.