Single threadded runtime
Forking/cloning from a multi threaded application does not work. After
a clone/fork only the calling thread is alive and if any of the other
threads allocated any resource in the c lib or kernel the clone can end
up in a deadlock. There's basically nothing we can do about that beside
all the advices from the net to *not* do anything crucial between the
clone/fork and the execve. Since the major thing of the runtime is to
adjust the environment of the child process prior to the exeve call this
is nearly impossible. Stress tests with the `tools/stress` utility show
that after a couple of shorthands of operations something fails.

In the northstar runtime there's no real need for multi threading since
almost everything is tied to the borrow of `State`. Everything that is
triggered by a "event" is handled in that context and not parallel. The
only two things where multi threading give any benefit is

* parallel bulk mounts (long time spent in kernel for each mount)
* handling of container io (stdout, stderr)

The mount are handled in thread pool from the futures crate and are
guaranteed to block and wait for anything when a container is cloned.
This could be done even more safe by tearing down the pool after a mount
operation that requires the pool (or has a speed benefit from the pool).

With the single thread approach the stop of containers needed to be
refactored because signaling the process and waiting for its termination
doesn't work. Instead of having the `stop` as a pair of `kill` and
`wait` it's split into `kill` and `wait` from the user perspective and
is reflected in the model for the console.