Setup SIGCHLD stream before spawning waitpid task
If the sigchild stream is established in the waitpid task the
application can be started and exit before the task is scheduled the
first time. The result is that the sigchld stream is setup *after* the
SIGCHLD of the container is set up and the event is lost.

Creating the stream outside of the waitpid task ensures that the start
signal from the runtime to the container is sent *after* the stream is
established.