Use random dm device names and remove wait for removal
Waiting for a dm device removal is contraproductive since this will
prevent the kernel from removing the dm device. The inotify handle makes
the device "in use" and without this patch the wait for removal will
wait forever since it's kind of self referencing.

Replace the use of heavy weight uuids with nanoid.