Do not run for Dependabot push events
For security reasons, Dependabot-triggered workflows run with
restricted, read-only access to the repository. This prevents the Clippy
action used from modifying the PR to add Clippy annotations to the
commits.

The Clippy action falls back appropriately if it detects that
GITHUB_HEAD_REF is set, but it is only set for workflows triggered by
the pull_request event. Because Dependabot pushes commits to a new
branch, it triggers both the push and pull_request events, so the
pull_request event succeeds thanks to the fallback, but the push event
does not.

For now (and maybe forever), fix that by not running the workflow on
push events that come from a Dependabot-created branch.

Related issue: actions-rs/clippy-check#2