Multi and parallel umounts
Extend the umounting code to leverage the same mechanism and the mount:
spawn a thraed per (blocking) umount. Mounts and umounts are not cpu
intensive but long lasting in the kernel. Extend the cli api with
multiple containers per umount.

Add nstar cli improvements.

Add a repository `mount` flag that instructs the runtime to mount *all*
containers of a repository upon runtime start.