Fix install to unknown repository
If a install request fails, the NPK needs to be drained from the
connection before continuing to process messages from a client. If the
runtime would break the reception upon any install error the next couple
of bytes read from the connections are very likely NPK (zip bytes) which
leads to a parsing error because the runtime tries to decode the data as
json.