Console extensions (#781)
* Extra permission to start a container with args and env

Starting a container with runtime defined arguments and env can be a
security thread. Keep the feature but hide it behind a permission that
must be carefully set.

* Move debug console config to debug section

Underline the purpose of the debug console by moving it to the debug
section of the runtime configuration. Extend documentation.
Remove the ability to configure permissions on the debug console.