Container IO refactoring (#777)
* Refactor io settings

* Remove fork wrapper and simplify socketpair usage

Remove the fork(2) wrapper because it works against the borrow checker
that is aware of branches that never return. This allow the safe close
(drop) of owned file descriptor handles like UnixStream.

* Never expose stdout/stderr of the runtime directly to containers

Instead of just dupping stdout/stderr, forward from a unix socket that
is passed into containers to stdout/stderr.

* Handle container creation errors

Kill any init that sends invalid data. Ensure that init is started
*after* the forker for sure received the pid of init.

* Update container start sequence

* Encapsulate runtime <-> forker channel

* Optimize file descriptor transfer

Use a single control message with and without a console fd passed to the
containers init.